Fix curl type warnings: use long instead of int
Multiple warnings appear during compilation of modules using libcurl:

```
gcc -fPIC -DPIC -O2 -flto=auto -ffat-lto-objects -fexceptions -g -grecord-gcc-switches -pipe -Wall -Wno-complain-wrong-lang -Werror=format-security -Wp,-U_FORTIFY_SOURCE,-D_FORTIFY_SOURCE=3 -Wp,-D_GLIBCXX_ASSERTIONS -specs=/usr/lib/rpm/redhat/redhat-hardened-cc1 -fstack-protector-strong -specs=/usr/lib/rpm/redhat/redhat-annobin-cc1 -m64 -march=x86-64 -mtune=generic -fasynchronous-unwind-tables -fstack-clash-protection -fcf-protection -mtls-dialect=gnu2 -fno-omit-frame-pointer -mno-omit-leaf-frame-pointer -DMOD_NAME='rest_client' -DPKG_MALLOC  -DSHM_MMAP  -DUSE_MCAST  -DDISABLE_NAGLE  -DSTATISTICS  -DHAVE_RESOLV_RES  -DF_MALLOC  -DQ_MALLOC  -DHP_MALLOC  -DDBG_MALLOC  -DF_PARALLEL_MALLOC  -DHAVE_STDATOMIC -DHAVE_GENERICS  -DNAME='"opensips"' -DVERSION='"3.6.2"' -DARCH='"x86_64"' -DOS='"linux"' -DCOMPILER='"gcc 15"' -D__CPU_x86_64 -D__OS_linux -D__SMP_yes -DCFG_DIR='"/etc/opensips/"'  -DVERSIONTYPE='"git"' -DTHISREVISION='"994bcd690"' -DFAST_LOCK -DADAPTIVE_WAIT -DADAPTIVE_WAIT_LOOPS=1024 -DHAVE_GETHOSTBYNAME2 -DHAVE_UNION_SEMUN -DHAVE_MSG_NOSIGNAL -DHAVE_MSGHDR_MSG_CONTROL -DHAVE_ALLOCA_H -DHAVE_TIMEGM -DHAVE_EPOLL -DHAVE_SIGIO_RT -DHAVE_SELECT -c rest_methods.c -o rest_methods.o
rest_methods.c: In function ‘init_transfer’:
rest_methods.c:408:17: warning: call to ‘Wcurl_easy_setopt_err_long’ declared with attribute warning: curl_easy_setopt expects a long argument [-Wattribute-warning]
  408 |                 w_curl_easy_setopt(handle, CURLOPT_HTTP_VERSION, curl_http_version);
      |                 ^
rest_methods.c:422:9: warning: call to ‘Wcurl_easy_setopt_err_long’ declared with attribute warning: curl_easy_setopt expects a long argument [-Wattribute-warning]
  422 |         w_curl_easy_setopt(handle, CURLOPT_VERBOSE, 1);
      |         ^
rest_methods.c:424:9: warning: call to ‘Wcurl_easy_setopt_err_long’ declared with attribute warning: curl_easy_setopt expects a long argument [-Wattribute-warning]
  424 |         w_curl_easy_setopt(handle, CURLOPT_FAILONERROR, 0);
      |         ^
rest_methods.c: In function ‘set_upload_opts’:
rest_methods.c:480:9: warning: call to ‘Wcurl_easy_setopt_err_long’ declared with attribute warning: curl_easy_setopt expects a long argument [-Wattribute-warning]
  480 |         w_curl_easy_setopt(handle, CURLOPT_POSTFIELDSIZE, body->len);
      |         ^
rest_methods.c: In function ‘rest_sync_transfer’:
rest_methods.c:714:17: warning: call to ‘Wcurl_easy_setopt_err_long’ declared with attribute warning: curl_easy_setopt expects a long argument [-Wattribute-warning]
  714 |                 set_post_opts(sync_handle, ctype, body);
      |                 ^
rest_methods.c: In function ‘start_async_http_req’:
rest_methods.c:847:17: warning: call to ‘Wcurl_easy_setopt_err_long’ declared with attribute warning: curl_easy_setopt expects a long argument [-Wattribute-warning]
  847 |                 set_post_opts(handle, req_ctype, req_body);
      |                 ^
rest_methods.c:918:25: warning: call to ‘Wcurl_easy_getinfo_err_long’ declared with attribute warning: curl_easy_getinfo expects a pointer to long [-Wattribute-warning]
  918 |                         curl_easy_getinfo(handle, CURLINFO_RESPONSE_CODE, &http_rc);
      |                         ^
```

Let's ensure we pass a literals and variables with a proper type.

(cherry picked from commit 2a03d51f7452b78e603ffb42bb32c815e27c078a)

+++ modules/db_http/http_dbase.c
@@ -804,7 +804,7 @@ int do_http_op (  const db_con_t* h, const db_key_t* k, const db_op_t* op,
 
 	LM_DBG("Sent:%s \n",q.s);
 
-	curl_easy_setopt(conn->handle, CURLOPT_HTTPGET, 1);
+	curl_easy_setopt(conn->handle, CURLOPT_HTTPGET, 1L);
 	curl_easy_setopt(conn->handle, CURLOPT_URL, q.s);
 
 
@@ -819,7 +819,7 @@ int do_http_op (  const db_con_t* h, const db_key_t* k, const db_op_t* op,
 	}
 
 
-	curl_easy_setopt(conn->handle, CURLOPT_FAILONERROR,1);
+	curl_easy_setopt(conn->handle, CURLOPT_FAILONERROR,1L);
 	ret = curl_easy_perform(conn->handle);
 
 
@@ -1054,8 +1054,8 @@ db_con_t* db_http_init(const str* url)
 		curl_easy_setopt(curl->handle,CURLOPT_HTTPHEADER,curl->headers);
 	}
 
-	curl_easy_setopt(curl->handle,CURLOPT_SSL_VERIFYPEER,0);
-	curl_easy_setopt(curl->handle,CURLOPT_SSL_VERIFYHOST,0);
+	curl_easy_setopt(curl->handle,CURLOPT_SSL_VERIFYPEER,0L);
+	curl_easy_setopt(curl->handle,CURLOPT_SSL_VERIFYHOST,0L);
 
 	curl_easy_setopt(curl->handle,CURLOPT_USERPWD,user_pass);
 	curl_easy_setopt(curl->handle,CURLOPT_HTTPAUTH,CURLAUTH_ANY);
@@ -1063,7 +1063,7 @@ db_con_t* db_http_init(const str* url)
 	curl_easy_setopt(curl->handle,CURLOPT_ERRORBUFFER,error_buffer);
 #if LIBCURL_VERSION_NUM >= 0x071002
 	LM_DBG("timeout set to %d", db_http_timeout);
-	curl_easy_setopt(curl->handle,CURLOPT_TIMEOUT_MS,db_http_timeout);
+	curl_easy_setopt(curl->handle,CURLOPT_TIMEOUT_MS,(long)db_http_timeout);
 #endif
 
 	ret = snprintf(path, DB_HTTP_BUFF_SIZE, "http");
